re-organize the interfaces around BaseTile.
